Fix comment issues for swagger reviews
