+++ src/dotnet/APIView/APIViewWeb/Client/src/shared/helpers.ts
@@ -30,3 +30,30 @@ export function updatePageSettings(callBack) {
     url: uri
   }).done(callBack());
 }
+
+export function getCodeRow(id: string) {
+  return $(`.code-line[data-line-id='${id}']`);
+}
+
+export function getCodeRowSectionClasses(id: string) {
+  var codeRow = getCodeRow(id);
+  var rowSectionClasses = "";
+  if (codeRow) {
+    rowSectionClasses = getRowSectionClasses(codeRow[0].classList);
+  }
+  return rowSectionClasses;
+}
+
+export function getRowSectionClasses(classList: DOMTokenList) {
+  const rowSectionClasses: string[] = [];
+  for (const value of classList.values()) {
+    if (value == "section-loaded" || value.startsWith("code-line-section-content") || value.match(/lvl_[0-9]+_(parent|child)_[0-9]+/)) {
+      rowSectionClasses.push(value);
+    }
+  }
+  return rowSectionClasses.join(' ');
+}
+
+export function toggleCommentIcon(id, show: boolean) {
+  getCodeRow(id).find(".icon-comments").toggleClass("invisible", !show);
+}
